Rename from TEAL debugger to AVM debugger & better organize deps
