+++ package.json
@@ -1,11 +1,11 @@
 {
-  "name": "teal-debug",
-  "displayName": "TEAL Debug",
-  "version": "0.52.0",
+  "name": "avm-debug",
+  "displayName": "AVM Debug",
+  "version": "1.0.0",
   "publisher": "algorand-vscode",
-  "description": "Debug extension for developing TEAL smart contracts in VS Code.",
+  "description": "Debug extension for developing AVM transactions and smart contracts in VS Code.",
   "author": {
-    "name": "Scytalians"
+    "name": "Algorand, llc"
   },
   "license": "MIT",
   "keywords": [
@@ -20,10 +20,10 @@
   "private": true,
   "repository": {
     "type": "git",
-    "url": "https://github.com/algorand/teal-debugger.git"
+    "url": "https://github.com/algorand/avm-debugger.git"
   },
   "bugs": {
-    "url": "https://github.com/algorand/teal-debugger/issues"
+    "url": "https://github.com/algorand/avm-debugger/issues"
   },
   "scripts": {
     "compile": "tsc -p ./",
@@ -70,10 +70,8 @@
     "@types/vscode": "^1.66.0",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "@vscode/debugadapter": "^1.64.0",
     "@vscode/debugadapter-testsupport": "^1.64.0",
     "@vscode/vsce": "^2.22.0",
-    "await-notify": "^1.0.1",
     "esbuild": "^0.14.29",
     "eslint": "^8.52.0",
     "eslint-config-prettier": "^9.0.0",
@@ -92,8 +90,8 @@
   "activationEvents": [
     "workspaceContains:**/*.teal",
     "onDebug",
-    "onDebugResolve:teal",
-    "onDebugDynamicConfigurations:teal"
+    "onDebugResolve:avm",
+    "onDebugDynamicConfigurations:avm"
   ],
   "workspaceTrust": {
     "request": "never"
@@ -122,11 +120,11 @@
     ],
     "debuggers": [
       {
-        "type": "teal",
+        "type": "avm",
         "languages": [
           "teal"
         ],
-        "label": "TEAL Debug",
+        "label": "AVM Debug",
         "program": "./out/src/debugAdapter/cli.js",
         "runtime": "node",
         "configurationAttributes": {
@@ -152,7 +150,7 @@
         },
         "initialConfigurations": [
           {
-            "type": "teal",
+            "type": "avm",
             "request": "launch",
             "name": "Debug AVM Transactions",
             "simulateTraceFile": "${workspaceFolder}/path/to/simulateTraceFile.json",
@@ -165,7 +163,7 @@
             "label": "AVM Debug",
             "description": "A new configuration for replaying and debugging a Algorand transactions.",
             "body": {
-              "type": "teal",
+              "type": "avm",
               "request": "launch",
               "name": "Debug AVM Transactions",
               "simulateTraceFile": "^\"\\${workspaceFolder}/path/to/simulateTraceFile.json\"",
@@ -178,7 +176,9 @@
     ]
   },
   "dependencies": {
+    "@vscode/debugadapter": "^1.64.0",
     "algosdk": "github:jasonpaulos/js-algorand-sdk#teal-source-map-improvements",
+    "await-notify": "^1.0.1",
     "json-bigint": "^1.0.0"
   }
 }
