+++ readme.md
@@ -1,16 +1,16 @@
-# TEAL Debug Adapter for VS Code
+# AVM Debugger
 
 ## Summary
 
-This repo contains a TEAL debugger which adheres to the [Debug Adapter Protocol](https://microsoft.github.io/debug-adapter-protocol/).
+This repo contains an AVM debugger which adheres to the [Debug Adapter Protocol](https://microsoft.github.io/debug-adapter-protocol/).
 This protocol is used by a variety of clients, and this repo specifically
-implements a client for VS Code.
+implements a basic client for VS Code.
 
 Unlike traditional debuggers, which typically execute code as it is being
 debugged, this debugger operates on an execution trace. The trace is created by
 the [algod simulate API](https://developer.algorand.org/docs/rest-apis/algod/#post-v2transactionssimulate).
 This debugger is not responsible for compiling programs, assembling transaction groups, or executing
-transactions/programs. It is only responsible for debugging the execution trace, which must already
+transactions/programs. It is only responsible for replaying the execution trace, which must already
 exist.
 
 This code is based on the [`vscode-mock-debug`](https://github.com/microsoft/vscode-mock-debug) repo.
