+++ src/web-extension.ts
@@ -1,10 +1,10 @@
 import * as vscode from 'vscode';
-import { activateTealDebug } from './activateMockDebug';
+import { activateAvmDebug } from './activateAvmDebug';
 import { InlineDebugAdapterFactory } from './internalDescriptorFactory';
 
 export function activate(context: vscode.ExtensionContext) {
   // Inline is the only supported mode for running the debug adapter in the browser.
-  activateTealDebug(context, new InlineDebugAdapterFactory());
+  activateAvmDebug(context, new InlineDebugAdapterFactory());
 }
 
 export function deactivate() {
